safari-534.53-branch/LayoutTests/fast/js/reentrant-call-unwind-expected.txt

This test checks that unwinding of exceptions properly copies registers of activation frames created by reentrant calls to JavaScript.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS f() is 0
PASS successfullyParsed is true

TEST COMPLETE
